Correctly depicted default gui elements to tell them apart
