+++ src/main/java/de/fabilucius/advancedperks/core/guisystem/element/defaultelements/CloseGuiWindowElement.java
@@ -1,7 +1,6 @@
-package de.fabilucius.advancedperks.core.guisystem.element.types;
+package de.fabilucius.advancedperks.core.guisystem.element.defaultelements;
 
 import de.fabilucius.advancedperks.core.guisystem.GuiSound;
-import de.fabilucius.advancedperks.core.guisystem.element.AbstractGuiElement;
 import de.fabilucius.advancedperks.core.guisystem.element.GuiElement;
 import de.fabilucius.advancedperks.core.guisystem.window.GuiWindow;
 import de.fabilucius.advancedperks.core.itembuilder.types.ItemStackBuilder;
@@ -10,7 +9,7 @@ import org.bukkit.event.inventory.InventoryClickEvent;
 
 import java.util.function.BiConsumer;
 
-public class CloseGuiWindowElement extends AbstractGuiElement {
+public class CloseGuiWindowElement extends AbstractDefaultGuiElement {
 
     public CloseGuiWindowElement(GuiWindow guiWindow, String displayName) {
         super(guiWindow, ItemStackBuilder.fromMaterial(Material.BARRIER)
