+++ src/main/java/de/fabilucius/advancedperks/core/guisystem/element/defaultelements/NextPageElement.java
@@ -1,8 +1,7 @@
-package de.fabilucius.advancedperks.core.guisystem.element.types;
+package de.fabilucius.advancedperks.core.guisystem.element.defaultelements;
 
 import de.fabilucius.advancedperks.core.guisystem.GuiSound;
 import de.fabilucius.advancedperks.core.guisystem.HeadTexture;
-import de.fabilucius.advancedperks.core.guisystem.element.AbstractGuiElement;
 import de.fabilucius.advancedperks.core.guisystem.element.GuiElement;
 import de.fabilucius.advancedperks.core.guisystem.window.types.AbstractPageGuiWindow;
 import de.fabilucius.advancedperks.core.itembuilder.types.SkullStackBuilder;
@@ -11,23 +10,28 @@ import org.bukkit.event.inventory.InventoryClickEvent;
 
 import java.util.function.BiConsumer;
 
-public class PreviousPageElement extends AbstractGuiElement {
+public class NextPageElement extends AbstractDefaultGuiElement {
 
     private final AbstractPageGuiWindow abstractPageGuiWindow;
+    private final int maxPage;
 
-    public PreviousPageElement(AbstractPageGuiWindow guiWindow, String displayName) {
+    public NextPageElement(AbstractPageGuiWindow guiWindow, String displayName, int maxPage) {
         super(guiWindow, SkullStackBuilder.fromMaterial(Material.PLAYER_HEAD)
-                .setBase64Value(HeadTexture.ARROW_LEFT.getValue())
+                .setBase64Value(HeadTexture.ARROW_RIGHT.getValue())
                 .setDisplayName(displayName)
                 .build());
         this.abstractPageGuiWindow = guiWindow;
+        this.maxPage = maxPage;
     }
 
     @Override
     public BiConsumer<GuiElement, InventoryClickEvent> handleInventoryClick() {
         return (guiElement, event) -> {
             event.setCancelled(true);
-            this.abstractPageGuiWindow.previousPage();
+            if (this.abstractPageGuiWindow.getPage() == this.maxPage) {
+                return;
+            }
+            this.abstractPageGuiWindow.nextPage();
             this.playSound(GuiSound.NORMAL_CLICK);
         };
     }
