+++ src/main/java/de/fabilucius/advancedperks/core/guisystem/element/defaultelements/PreviousPageElement.java
@@ -1,8 +1,7 @@
-package de.fabilucius.advancedperks.core.guisystem.element.types;
+package de.fabilucius.advancedperks.core.guisystem.element.defaultelements;
 
 import de.fabilucius.advancedperks.core.guisystem.GuiSound;
 import de.fabilucius.advancedperks.core.guisystem.HeadTexture;
-import de.fabilucius.advancedperks.core.guisystem.element.AbstractGuiElement;
 import de.fabilucius.advancedperks.core.guisystem.element.GuiElement;
 import de.fabilucius.advancedperks.core.guisystem.window.types.AbstractPageGuiWindow;
 import de.fabilucius.advancedperks.core.itembuilder.types.SkullStackBuilder;
@@ -11,28 +10,23 @@ import org.bukkit.event.inventory.InventoryClickEvent;
 
 import java.util.function.BiConsumer;
 
-public class NextPageElement extends AbstractGuiElement {
+public class PreviousPageElement extends AbstractDefaultGuiElement {
 
     private final AbstractPageGuiWindow abstractPageGuiWindow;
-    private final int maxPage;
 
-    public NextPageElement(AbstractPageGuiWindow guiWindow, String displayName, int maxPage) {
+    public PreviousPageElement(AbstractPageGuiWindow guiWindow, String displayName) {
         super(guiWindow, SkullStackBuilder.fromMaterial(Material.PLAYER_HEAD)
-                .setBase64Value(HeadTexture.ARROW_RIGHT.getValue())
+                .setBase64Value(HeadTexture.ARROW_LEFT.getValue())
                 .setDisplayName(displayName)
                 .build());
         this.abstractPageGuiWindow = guiWindow;
-        this.maxPage = maxPage;
     }
 
     @Override
     public BiConsumer<GuiElement, InventoryClickEvent> handleInventoryClick() {
         return (guiElement, event) -> {
             event.setCancelled(true);
-            if (this.abstractPageGuiWindow.getPage() == this.maxPage) {
-                return;
-            }
-            this.abstractPageGuiWindow.nextPage();
+            this.abstractPageGuiWindow.previousPage();
             this.playSound(GuiSound.NORMAL_CLICK);
         };
     }
