+++ src/main/java/de/fabilucius/advancedperks/guisystem/perkgui/PerkGuiWindow.java
@@ -5,10 +5,10 @@ import de.fabilucius.advancedperks.configuration.ConfigurationLoader;
 import de.fabilucius.advancedperks.configuration.exception.ConfigurationInitializationException;
 import de.fabilucius.advancedperks.core.MessagesConfiguration;
 import de.fabilucius.advancedperks.core.SettingsConfiguration;
-import de.fabilucius.advancedperks.core.guisystem.element.types.CloseGuiWindowElement;
-import de.fabilucius.advancedperks.core.guisystem.element.types.GuiBackgroundElement;
-import de.fabilucius.advancedperks.core.guisystem.element.types.NextPageElement;
-import de.fabilucius.advancedperks.core.guisystem.element.types.PreviousPageElement;
+import de.fabilucius.advancedperks.core.guisystem.element.defaultelements.CloseGuiWindowElement;
+import de.fabilucius.advancedperks.core.guisystem.element.defaultelements.GuiBackgroundElement;
+import de.fabilucius.advancedperks.core.guisystem.element.defaultelements.NextPageElement;
+import de.fabilucius.advancedperks.core.guisystem.element.defaultelements.PreviousPageElement;
 import de.fabilucius.advancedperks.core.guisystem.window.types.AbstractPageGuiWindow;
 import de.fabilucius.advancedperks.data.PerkData;
 import de.fabilucius.advancedperks.data.PerkDataRepository;
